+++ spring-kotlin-coroutine/src/main/kotlin/org/springframework/kotlin/experimental/coroutine/web/CoroutinesWebMvcConfigurerAdapter.kt
@@ -16,7 +16,6 @@
 
 package org.springframework.kotlin.experimental.coroutine.web
 
-import kotlinx.coroutines.experimental.Unconfined
 import org.springframework.beans.factory.config.BeanDefinition
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Role
@@ -34,6 +33,7 @@ import org.springframework.web.method.support.ModelAndViewContainer
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
 import kotlin.coroutines.experimental.Continuation
 import kotlin.coroutines.experimental.CoroutineContext
+import kotlin.coroutines.experimental.EmptyCoroutineContext
 import kotlin.coroutines.experimental.intrinsics.COROUTINE_SUSPENDED
 
 @Configuration
@@ -85,7 +85,7 @@ private class DeferredResultContinuation: Continuation<Any> {
     val deferredResult = DeferredResult<Any>()
 
     override val context: CoroutineContext
-        get() = Unconfined
+        get() = EmptyCoroutineContext
 
     override fun resume(value: Any) {
         deferredResult.setResult(value)